fetch categorised expenses

+++ core/src/main/kotlin/domfin/repository/tables/Transactions.kt
@@ -10,7 +10,7 @@ object Transactions : Table("transactions") {
     val status = text("status").check("transactions_check_status", { it inList listOf("booked", "pending") })
     val bookingDate = varchar("booking_date", 10)
     val valueDate = varchar("value_date", 10)
-    val amount = integer("amount")
+    val amount = long("amount")
     val currency = text("currency")
     val creditorName = text("creditor_name").nullable()
     val debtorName = text("debtor_name").nullable()
